\name{plot.corrMove}
\alias{plot.corrMove}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot corrMove objects
}
\description{
Generates a three-panel plot consisting of the three MCIs (etaDif, etaDft, and etaTot) with confidence intervals vs. time. Within each panel, different plot segments depict the different partitions in the data.
}
\usage{
plot(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A corrMove object, generated by the corrMove function.
}
  \item{\dots}{
Additional options supplied to plot.
}
}
\details{
Colors correspond to the selected model for each partition of the dataset as: UU model=green, CU model=orange, UC model=purple, and CC model=pink.
}
%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
Justin M. Calabrese
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.

%## The function is currently defined as
%function (corrMove, ...)
%{
%    buf <- 0.0125
%    with(corrMove, {
%        minVal <- min(c(min(etaDif.CI.Low), min(etaDft.CI.Low),
%            min(etaTot.CI.Low)))
%        maxVal <- max(c(max(etaDif.CI.Upp), max(etaDft.CI.Upp),
%            max(etaTot.CI.Upp)))
%        yMin <- minVal - buf
%        yMax <- maxVal + buf
%        cols <- c("#1b9e77", "#d95f02", "#7570b3", "#e7298a")
%        m <- matrix(c(1, 2, 3, 4), nrow = 4, ncol = 1, byrow = TRUE)
%        layout(mat = m, heights = c(0.3, 0.3, 0.3, 0.1))
%        par(mar = c(2.5, 5, 1, 1))
%        plot(timestamp, etaDif.MLE, pch = 20, col = cols[sel.model.num],
%            ylim = c(yMin, yMax), ylab = "Diffusive corr.", xlab = "",
%            cex.lab = 1.5, cex.axis = 1.5)
%        points(timestamp, etaDif.CI.Low, pch = "-", col = cols[sel.model.num])
%        points(timestamp, etaDif.CI.Upp, pch = "-", col = cols[sel.model.num])
%        plot(timestamp, etaDft.MLE, pch = 20, col = cols[sel.model.num],
%            ylim = c(yMin, yMax), ylab = "Drift corr.", xlab = "",
%            cex.lab = 1.5, cex.axis = 1.5)
%        points(timestamp, etaDft.CI.Low, pch = "-", col = cols[sel.model.num])
%        points(timestamp, etaDft.CI.Upp, pch = "-", col = cols[sel.model.num])
%        plot(timestamp, etaTot.MLE, pch = 20, col = cols[sel.model.num],
%            ylim = c(yMin, yMax), ylab = "Total corr.", xlab = "",
%            cex.lab = 1.5, cex.axis = 1.5)
%        points(timestamp, etaTot.CI.Low, pch = "-", col = cols[sel.model.num])
%        points(timestamp, etaTot.CI.Upp, pch = "-", col = cols[sel.model.num])
%        modNums <- unique(sel.model.num)
%        modCodes <- as.character(unique(sel.mod.code))
%        par(mar = c(0, 5, 0, 1))
%        plot(1, type = "n", axes = FALSE, xlab = "", ylab = "")
%        legend("center", legend = modCodes, fill = cols[modNums],
%            ncol = length(modNums), cex = 1.75)
%    })
%  }
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
